updates the variable type for the string len

+++ ex05/ft_print_comb4.c
@@ -6,7 +6,7 @@
 /*   By: argrouss <argrouss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/08 19:29:37 by argrouss          #+#    #+#             */
-/*   Updated: 2024/02/09 01:38:43 by argrouss         ###   ########.fr       */
+/*   Updated: 2024/02/09 02:38:56 by argrouss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #include <stdio.h>
 
 void	ft_print_comb(void);
-void	check_to_print(char *nb_to_check, int nb_to_check_size);
+void	check_to_print(char *nb_to_check, size_t nb_to_check_size);
 
 int	main(void)
 {
@@ -25,7 +25,7 @@ int	main(void)
 void	ft_print_comb(void)
 {
 	char	nb_f[6];
-	int		nb_f_size;
+	size_t	nb_f_size;
 
 	nb_f[0] = 48;
 	nb_f[3] = 44;
@@ -49,7 +49,7 @@ void	ft_print_comb(void)
 	}
 }
 
-void	check_to_print(char *nb_to_check, int nb_to_check_size)
+void	check_to_print(char *nb_to_check, size_t nb_to_check_size)
 {
 	if (nb_to_check[0] < nb_to_check[1] && (nb_to_check[1] < nb_to_check[2]))
 	{
